pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "argus-lite"
version = "1.0.0"
description = "Local security scanner for authorized penetration testing"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "python-dateutil>=2.8",
    "rich>=13.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
]

[project.scripts]
argus-lite = "argus_lite.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.coverage.run]
source = ["argus_lite"]

[tool.coverage.report]
fail_under = 70
